// GoTaskServicePlus/GoTaskServiceplus.Client/src/app/Models/Segurity/Login/LoginModel.ts
export class Login {

  Password: string;
  Email: string;
  constructor() {

  }
}

export class LoginRegister {

  Password1: string;
  Password2: string;
  Email: string;
  NumberWhatsapp: string;
 
  constructor() {

  }
}


export class StatusLogim {

  token: string;
  dateStart: string;
  dateEnd: string;
  expiteTimeMinute: number;


}

export class ChangePassword {

  password: string;
  email: string;
  code: string;
 
}